All fixed

+++ BeatMyBot/Networking.h
@@ -42,7 +42,7 @@ private:
   {
     int16_t xPos;   // X value of the bots position, shortened to 16 bit int
     int16_t yPos;   // Y value of the bots position, shortened to 16 bit int
-    int16_t dir;     // Degrees used so int can be used with more effectiveness
+    float dir;     // Degrees used so int can be used with more effectiveness
     float xVel;
     float yVel;
     bool isAlive;     // If the bot is alive or not
